Create phoneDictionary.cpp
This piece of code basically simulates an interview problem where Name and phone numbers were inserted into the dictionary. And the searching was done based on the phone number. 

PS: Deletion was mentioned in the interview. If it was present then we need to have a map instead of vectors.